Allow configuration of per-broker concurrency.
Broker requests are stored in a channel while waiting for replies.
Currently this channel has a fixed size of 4, limiting the number of
concurrent requests per-broker to 4. This causes problems when listening
on many topic-partitions concurrently.

This commit allows configuration of this buffer size via a new
`ConcurrencyPerBroker` parameter in `ClientConfig`.