JVM-compatible hash partitioner
Proof-of-concept; it may or may not be worth actually merging this (not sure if
this counts as an API change or not since it will change partitioning for people
who upgrade).

Also, I haven't actually tested to see if it gives the same value or not,
though it ought to.