prod: drain successes on close just like errors
We kindly drain the errors channel on close so that users who stop reading from
it when calling close (e.g. in a defer) don't hang. We should do the same thing
for the successes channel when AckSuccesses is defined.

May or may not fix the bug reported tangentially at the bottom of
https://github.com/Shopify/sarama/issues/180#issuecomment-63127314
where shutting down with AckSuccesses could be very slow.