client: close backgroundMetadataUpdater faster
Rather than waiting for the next timer cycle (defaults to ten minutes) have it
read from two channels in a select so it tears down right away. It was polluting
test backtraces (where we create and destroy a lot of clients) and making
debugging more difficult than it needed to be.