Change multi-consumer unexpected error handling
Previously on unexpected errors (i.e. missing messages in the response) we would
bubble the error to the user and otherwise carry on. These kind of errors should
only happen when talking to misbehaving brokers, and while an argument can be
made that we should just abort that broker connection on such an error, we
didn't want to disrupt working connections either if it was just the one
topic/partition that was misbehaving.

In this version, we redispatch the partition as well. This has a couple of
advantages:
- if the broker itself is bad for some reason, we'll end up redispatching *all*
  the partitions assigned to it, which should implicitly bounce the broker
  connection
- if the problem is out-of-sync metadata or leadership in the broker, this gives
  the partition a chance to end up on some other more sane leader

In the case of a truly broken cluster, we'll end up bouncing between metadata
updates and fetches, which probably won't be a big deal.