Use Metadata field in SimpleProducer
Rather than storing a `map` from messages to expectations, store the
expectations in our handy new `Metadata` field. This is certainly simpler, and
is probably slightly more efficient.

Also call `AsyncClose` instead of `Close` when shutting down the underlying
Producer, to avoid the hypothetical race condition where `Close` swallows a
success or error and that call to `SendMessage` never returns.