Don't pre-emptively test the leader connectedness
Write a test, find a bug (well, not really).

Before multiple-retries were in place, I'd added this pre-emptive check for
leader connectedness in the leaderDispatcher so that if we got back stale
metadata we didn't bother going through the whole retry rigamarole and failing
anyways. Now that we have multiple-retries, this check is actually causing us to
short-circuit our retries in one case, which isn't nice. Since it isn't
necessary anymore, just remove it.

Discovered because the test I wrote for stale metadata tried to make more
metadata requests than I was expecting, and subsequently hung.