client: connect to a random seed broker
Rather than walking them in order, walk them randomly using go's map iterator.
This should lighten the load on the first seed in a lot of failover scenarios.

Also rework a bunch of how seeds are stored:
- store both seeds and dead brokers as Broker objects, not addresses, to avoid
  constructing new Brokers all the time
- lazily set `seedBroker` in `any()` rather than eagerly in the constructor and
  in `disconnectBroker()`
- never add non-seed brokers to `deadBrokers`
- substantially simplify `resurrectDeadBrokers`

This doesn't break the predictability of any of our tests, fortunately, because
we only ever pass one seed in anyways.