Get rid of `pending` in the brokerProducer
That was a bad pattern because it could easily have gotten out-of-sync with what
was actually sent to the flusher due to subtle bugs. Instead, have the flusher
pass back the set it sent alongside the response. This forces us to use a
struct, which lets us combine the `responses` and `errors` channels too.