This refactor just got a lot harder to swallow
The problem is that receiving a response can change the retry state of a
partition (or the entire broker). This potentially requires us to retry any
messages we'd had buffered for that partition, which changes the size of the
produceSet we'd accumulated, potentially even making it empty again. Things are
complicated further by the fact that we must be prepared to handle responses
while forcing blocking flushes, midway through the message processing.

This does a quick fix, but I'm not really happy with it:
- Flush the buffer in the appropriate response/error cases.
- Move the blocking flush to the beginning of the message processing so any
  state changes it causes are picked up correctly for that message.
- When we handle a response which leaves our buffer empty, don't try and send
  it. This isn't perfect in that a response which shrinks our buffer below the
  flush threshold may still end up getting flushed, but at least we won't flush
  a buffer that's entirely empty.