Refactor brokerProducer again
Turn the "if wouldOverflow then flush" logic into a more complete `waitForSpace`
method which handles all of that logic including re-checking error states and
buffer sizes as necessary.

Extract the "when should we be setting `output`" logic into the end of every
single iteration. This requires tracking an additional `timerFired` instance
variable on the `brokerProducer`, but is much easier to follow and allows us
to be more precise when e.g. a response flushes part but not all of our buffer.

Extract the "if this is a chaser, change state" logic out of `retryReason` and
rename it to `needsRetry`. This makes it a pure function and avoids unnecessary
work in `waitForSpace`.