Make the response ReadTimeout start on send
Currently the broker sets the ReadTimeout from the moment when it starts
listening for the response. However, due to the serial blocking nature of kafka
network connections, this can be substantially later than the time we actually
sent the request, and can lead to long useless waits when the connection has
timed out but there are requests in the queue.

Instead, set the ReadTimeout from the moment we have successfully sent the
request.