Send messages and errors directly in consumer mock
The mock PartitionConsumer sent messages to an expectations channel
on calls to YieldMessage and YieldError. This led to race conditions
in tests if pc.Messages was called before the handleExpectations
goroutine added a message to the channel.

Messages and errors are now sent to their respective channels
directly, and the expectations channel & goroutine are removed
altogether.