Minor lint changes
Partly cherry-picked from #575:
- remove an unnecessary `_`
- fix a few doc comments
- reorder an import
- elide nested slice literal types

The last one bears a bit of explanation: if you have a nested literal of a
complex type, you can ellide the inner literal types because they can be implied
from the outermost declaration. You've been able to do this for inner slices
since Go 1, but they only added the ability to do it for maps starting in Go 1.5
so we only do it for slices for now.