Fix race condition on OffsetManager shutdown
If `updateCommitted` ran between the `Unlock` call and the read from `clean`
then we might miss the transition message and hang. Instead, read
unconditionally from `clean` and launch a second goroutine that *writes* a
message to it if we were already clean.

The additional local `cancel` is necessary to avoid leaking the second (writing)
goroutine in the event that the main shutdown goroutine picks up a real
transitional event before our extra writer even runs; without it the writer
could block writing forever.