Adding Dockerfile for easily bringing up tools
This provides an easy way to run the producer and consumer using docker containers.

The following examples use kafka in a docker image as the broker, and therefore
have a link to the kafka container to use as the hostname.  This will work the same
as long as kafka is reachable by the docker container.

Example build:
    docker build -t sarama .

Example producer (this example assumes kafka running as another container):
    docker run --rm --link kafka:kafka sarama kafka-console-producer -topic test -brokers kafka:9092 -value "test 1 2"

Example consumer
    docker run --rm --link kafka:kafka sarama kafka-console-consumer -topic test -offset oldest -brokers kafka:9092

Reported by: Tyler Ruppert <tylerruppert@gmail.com>
Signed-off-by: Tyler Ruppert <tylerruppert@gmail.com>