Add a config option to store only minimal metadata
Typically the default is just to fetch and store all topics, as it is
simpler and more convenient not to have to go to the network for every
new topic that is needed. However this can use a lot of memory for
clusters with many topics and partitions, so add a flag which can be
unset in order to just store the minimal topic information needed so
far.