Make Partitioner.RequiresConsistency vary per-message
One of the useful features of the random partitioner is that it will
only partition messages onto partitions that are currently online. In
theory, hash partitioner is capable of exhibiting the same behavior for
messages where the key is unset, but it can't because
RequiresConsistency is a constant for the partitioner. It's also not
feasible to add methods to the Partitioner interface, since that's part
of the public API.

Therefore, this introduces a new interface, ExtendedPartitioner, which
includes an additional method: MessageRequiresConsistency. This lets the
partitioner make a per-message determination for whether consistency is
required (and therefore whether it's acceptable to publish to partitions
which are offline). The producer library will use this method if it's
available, and fall back on the existing RequiresConsistency method
otherwise.

Since hash partitioners are the only ones which have different
consistency requirements per-message, for now they are the only
partitioner to implement MessageRequiresConsistency.