bug: remove invalid test
I'm not sure what the purpose of
`TestClusterAdminCreateTopicWithDiffVersion` originally was, but rather
than testing desired behaviour, it seemed to be asserting that a
previous deficiency in the mockbroker responses would return an
ErrInsufficientData because a V1 request got a V0 response from the
mockbroker. Since ec2cbc3 the mockbroker responses with a matching
response version to the original request, so this test would no longer
pass its assertions and should be removed.