Fix possible faulty metrics in TestFuncProducing* (#1545)
Do not check the request-latency-in-ms metric for the cluster as it
could be 0 if one of the broker is down. Similarly the response-size
metric for the cluster could be recorded with 0 bytes.
The respective broker metrics on the other hand are useful.

Dump applicative logs from all brokers in case the tests (or something
other build task) fail.